requirements.txt
flask
flask-login
flask-sqlalchemy
werkzeug
email-validator
python-dotenv
pytest
numpy 